correction on file organisation
