+++ .gitignore
@@ -1,0 +1,20 @@
+#include <stdio.h>
+#include "main.h"
+int main(void)
+{
+	 int len;
+    int len2;
+    unsigned int ui;
+    void *addr;
+
+    len = _printf("Let's try to printf a simple sentence.\n");
+    len2 = printf("Let's try to printf a simple sentence.\n");
+    _printf("String:[%s]\n", "I am a string !");
+    printf("String:[%s]\n", "I am a string !");
+        len = _printf("Percent:[%%]\n");
+    len2 = printf("Percent:[%%]\n");
+    _printf("Character:[%c]\n", 'H');
+    printf("Character:[%c]\n", 'H');
+       _printf("Address:[%p]\n", addr);
+    printf("Address:[%p]\n", addr);
+}
