+++ _putchar.c
@@ -1,5 +1,5 @@
 #include <unistd.h>
-
+#include "main.h"
 /**
  * _putchar - writes the character c to stdout
  * @c: The character to print
